Move LowerBoundary implementation to Base2ExponentialBucketHistogram class
